pyspark/requirements.txt
pyspark=3.2.1
python-dotenv=0.21.0
